admin_portal/requirements.txt
asgiref==3.11.1
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
